backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pymupdf==1.24.0
python-dotenv==1.0.1
groq==0.9.0
openai==1.30.0
anthropic==0.28.0
pydantic==2.7.0